CustomFeatures : Split Custom Features in its own CustomFeatures.cpp/h files
