+++ es-app/src/CustomFeatures.cpp
@@ -1,0 +1,370 @@
+#include "CustomFeatures.h"
+
+#include "utils/FileSystemUtil.h"
+#include "utils/StringUtil.h"
+#include "Log.h"
+
+bool CustomFeatures::FeaturesLoaded = false;
+
+CustomFeatures CustomFeatures::GlobalFeatures;
+CustomFeatures CustomFeatures::SharedFeatures;
+std::map<std::string, EmulatorData> CustomFeatures::EmulatorFeatures;
+
+bool CustomFeatures::loadEsFeaturesFile()
+{
+	EmulatorFeatures.clear();
+	FeaturesLoaded = false;
+
+	GlobalFeatures.clear();
+	SharedFeatures.clear();
+
+	std::string path = Utils::FileSystem::getEsConfigPath() + "/es_features.cfg";
+	if (!Utils::FileSystem::exists(path))
+		path = Utils::FileSystem::getSharedConfigPath() + "/es_features.cfg";
+
+	if (!Utils::FileSystem::exists(path))
+		return false;
+
+	pugi::xml_document doc;
+	pugi::xml_parse_result res = doc.load_file(path.c_str());
+
+	if (!res)
+	{
+		LOG(LogError) << "Could not parse es_features.cfg file!";
+		LOG(LogError) << res.description();
+		return false;
+	}
+
+	pugi::xml_node systemList = doc.child("features");
+	if (!systemList)
+	{
+		LOG(LogError) << "es_features.cfg is missing the <features> tag!";
+		return false;
+	}
+
+	pugi::xml_node sharedFeatures = systemList.child("sharedFeatures");
+	if (sharedFeatures)
+		SharedFeatures = loadCustomFeatures(sharedFeatures);
+
+	pugi::xml_node globalFeatures = systemList.child("globalFeatures");
+	if (globalFeatures)
+		GlobalFeatures = loadCustomFeatures(globalFeatures);
+
+	FeaturesLoaded = true;
+
+	for (pugi::xml_node emulator = systemList.child("emulator"); emulator; emulator = emulator.next_sibling("emulator"))
+	{
+		if (!emulator.attribute("name"))
+			continue;
+
+		std::string emulatorNames = emulator.attribute("name").value();
+		for (auto tmpName : Utils::String::split(emulatorNames, ','))
+		{
+			std::string emulatorName = Utils::String::trim(tmpName);
+			if (EmulatorFeatures.find(emulatorName) == EmulatorFeatures.cend())
+			{
+				EmulatorData emul;
+				emul.name = emulatorName;
+				EmulatorFeatures[emulatorName] = emul;
+			}
+
+			EmulatorFeatures::Features emulatorFeatures = EmulatorFeatures::Features::none;
+
+			auto customEmulatorFeatures = loadCustomFeatures(emulator);
+
+			if (emulator.attribute("features") || customEmulatorFeatures.size() > 0)
+			{
+				if (emulator.attribute("features"))
+					emulatorFeatures = EmulatorFeatures::parseFeatures(emulator.attribute("features").value());
+
+				auto it = EmulatorFeatures.find(emulatorName);
+				if (it != EmulatorFeatures.cend())
+				{
+					auto& emul = it->second;
+					if (emul.name != emulatorName)
+						continue;
+
+					emul.features = (EmulatorFeatures::Features) (emul.features | emulatorFeatures);
+					for (auto feat : customEmulatorFeatures)
+						emul.customFeatures.push_back(feat);
+				}
+			}
+
+			pugi::xml_node coresNode = emulator.child("cores");
+			if (coresNode == nullptr)
+				coresNode = emulator;
+
+			for (pugi::xml_node coreNode = coresNode.child("core"); coreNode; coreNode = coreNode.next_sibling("core"))
+			{
+				if (!coreNode.attribute("name"))
+					continue;
+
+				std::string coreNames = coreNode.attribute("name").value();
+				for (auto tmpCoreName : Utils::String::split(coreNames, ','))
+				{
+					std::string coreName = Utils::String::trim(tmpCoreName);
+
+					EmulatorFeatures::Features coreFeatures = coreNode.attribute("features") ? EmulatorFeatures::parseFeatures(coreNode.attribute("features").value()) : EmulatorFeatures::Features::none;
+					auto customCoreFeatures = loadCustomFeatures(coreNode);
+
+					bool coreFound = false;
+
+					for (auto it = EmulatorFeatures.begin(); it != EmulatorFeatures.end(); it++)
+					{
+						auto& emul = it->second;
+						if (emul.name != emulatorName)
+							continue;
+
+						for (auto& core : emul.cores)
+						{
+							if (core.name == coreName)
+							{
+								coreFound = true;
+								core.features = (EmulatorFeatures::Features) (core.features | coreFeatures);
+
+								for (auto feat : customCoreFeatures)
+									core.customFeatures.push_back(feat);
+							}
+						}
+
+						if (!coreFound)
+						{
+							CoreData core;
+							core.name = coreName;
+							core.features = coreFeatures;
+							core.customFeatures = customCoreFeatures;
+							emul.cores.push_back(core);
+						}
+					}
+
+					pugi::xml_node systemsCoresNode = coreNode.child("systems");
+					if (systemsCoresNode == nullptr)
+						systemsCoresNode = coreNode;
+
+					for (pugi::xml_node systemNode = systemsCoresNode.child("system"); systemNode; systemNode = systemNode.next_sibling("system"))
+					{
+						if (!systemNode.attribute("name"))
+							continue;
+
+						std::string systemName = systemNode.attribute("name").value();
+
+						EmulatorFeatures::Features systemFeatures = systemNode.attribute("features") ? EmulatorFeatures::parseFeatures(systemNode.attribute("features").value()) : EmulatorFeatures::Features::none;
+						auto customSystemFeatures = loadCustomFeatures(systemNode);
+						if (systemFeatures == EmulatorFeatures::Features::none && customSystemFeatures.size() == 0)
+							continue;
+
+						for (auto it = EmulatorFeatures.begin(); it != EmulatorFeatures.end(); it++)
+						{
+							auto& emul = it->second;
+							if (emul.name != emulatorName)
+								continue;
+
+							for (auto& core : emul.cores)
+							{
+								if (core.name == coreName)
+								{
+									bool systemFound = false;
+
+									for (auto& systemFeature : core.systemFeatures)
+									{
+										if (systemFeature.name == systemName)
+										{
+											systemFound = true;
+											systemFeature.features = (EmulatorFeatures::Features) (systemFeature.features | systemFeatures);
+
+											for (auto feat : customSystemFeatures)
+												systemFeature.customFeatures.push_back(feat);
+										}
+									}
+
+									if (!systemFound)
+									{
+										SystemFeature sysF;
+										sysF.name = systemName;
+										sysF.features = systemFeatures;
+										sysF.customFeatures = customSystemFeatures;
+										core.systemFeatures.push_back(sysF);
+									}
+								}
+							}
+						}
+					}
+				}
+			}
+
+			pugi::xml_node systemsNode = emulator.child("systems");
+			if (systemsNode == nullptr)
+				systemsNode = emulator;
+
+			for (pugi::xml_node systemNode = systemsNode.child("system"); systemNode; systemNode = systemNode.next_sibling("system"))
+			{
+				if (!systemNode.attribute("name"))
+					continue;
+
+				std::string systemName = systemNode.attribute("name").value();
+
+				EmulatorFeatures::Features systemFeatures = systemNode.attribute("features") ? EmulatorFeatures::parseFeatures(systemNode.attribute("features").value()) : EmulatorFeatures::Features::none;
+				auto customSystemFeatures = loadCustomFeatures(systemNode);
+
+				auto it = EmulatorFeatures.find(emulatorName);
+				if (it != EmulatorFeatures.cend())
+				{
+					auto& emul = it->second;
+					bool systemFound = false;
+					for (auto& systemFeature : emul.systemFeatures)
+					{
+						if (systemFeature.name == systemName)
+						{
+							systemFound = true;
+							systemFeature.features = (EmulatorFeatures::Features) (systemFeature.features | systemFeatures);
+
+							for (auto feat : customSystemFeatures)
+								systemFeature.customFeatures.push_back(feat);
+						}
+					}
+
+					if (!systemFound)
+					{
+						SystemFeature sysF;
+						sysF.name = systemName;
+						sysF.features = systemFeatures;
+						sysF.customFeatures = customSystemFeatures;
+						emul.systemFeatures.push_back(sysF);
+					}
+				}
+			}
+		}
+	}
+
+	GlobalFeatures.sort();
+	return true;
+}
+
+
+CustomFeatures CustomFeatures::loadCustomFeatures(pugi::xml_node node)
+{
+	CustomFeatures ret;
+
+	if (node.attribute("features"))
+	{
+		auto features = node.attribute("features").value();
+		for (auto name : Utils::String::split(features, ','))
+		{
+			std::string featureValue = Utils::String::trim(name);
+			auto it = std::find_if(SharedFeatures.cbegin(), SharedFeatures.cend(), [featureValue](const CustomFeature& x) { return x.value == featureValue; });
+
+			if (it != SharedFeatures.cend())
+				ret.push_back(*it);
+		}
+	}
+
+	pugi::xml_node customFeatures = node.child("features");
+	if (customFeatures == nullptr)
+		customFeatures = node;
+
+	for (pugi::xml_node featureNode = customFeatures.first_child(); featureNode; featureNode = featureNode.next_sibling())
+	{
+		std::string name = featureNode.name();
+		if (name == "sharedFeature")
+		{
+			auto it = SharedFeatures.cend();
+
+			if (featureNode.attribute("name"))
+			{
+				std::string featureName = featureNode.attribute("name").value();
+				it = std::find_if(SharedFeatures.cbegin(), SharedFeatures.cend(), [featureName](const CustomFeature& x) { return x.name == featureName; });
+			}
+
+			if (it == SharedFeatures.cend() && featureNode.attribute("value"))
+			{
+				std::string featureValue = featureNode.attribute("value").value();
+				it = std::find_if(SharedFeatures.cbegin(), SharedFeatures.cend(), [featureValue](const CustomFeature& x) { return x.value == featureValue; });
+			}
+
+			if (it != SharedFeatures.cend())
+			{
+				CustomFeature cs = *it;
+
+				if (featureNode.attribute("submenu"))
+					cs.submenu = featureNode.attribute("submenu").value();
+
+				if (featureNode.attribute("preset"))
+					cs.preset = featureNode.attribute("preset").value();
+
+				if (featureNode.attribute("group"))
+					cs.group = featureNode.attribute("group").value();
+
+				if (featureNode.attribute("order"))
+					cs.order = Utils::String::toInteger(featureNode.attribute("order").value());
+
+				ret.push_back(cs);
+			}
+
+			continue;
+		}
+		else if (name != "feature")
+			continue;
+
+		if (!featureNode.attribute("name"))
+			continue;
+
+		CustomFeature feat;
+		feat.name = featureNode.attribute("name").value();
+
+		if (featureNode.attribute("description"))
+			feat.description = featureNode.attribute("description").value();
+
+		if (featureNode.attribute("submenu"))
+			feat.submenu = featureNode.attribute("submenu").value();
+
+		if (featureNode.attribute("preset"))
+			feat.preset = featureNode.attribute("preset").value();
+
+		if (featureNode.attribute("group"))
+			feat.group = featureNode.attribute("group").value();
+
+		if (featureNode.attribute("order"))
+			feat.order = Utils::String::toInteger(featureNode.attribute("order").value());
+
+		if (featureNode.attribute("value"))
+			feat.value = featureNode.attribute("value").value();
+		else
+			feat.value = Utils::String::replace(feat.name, " ", "_");
+
+		for (pugi::xml_node value = featureNode.child("choice"); value; value = value.next_sibling("choice"))
+		{
+			if (!value.attribute("name"))
+				continue;
+
+			CustomFeatureChoice cv;
+			cv.name = value.attribute("name").value();
+
+			if (value.attribute("value"))
+				cv.value = value.attribute("value").value();
+			else
+				cv.value = Utils::String::replace(cv.name, " ", "_");
+
+			feat.choices.push_back(cv);
+		}
+
+		if (feat.choices.size() > 0 || !feat.preset.empty())
+			ret.push_back(feat);
+	}
+
+	return ret;
+}
+
+void CustomFeatures::sort()
+{
+	std::sort(begin(), end(), [](CustomFeature& feat1, CustomFeature& feat2) { return feat1.order < feat2.order; });
+}
+
+bool CustomFeatures::hasFeature(const std::string& name) const
+{
+	return any([name](auto x) { return x.value == name; });
+}
+
+bool CustomFeatures::hasGlobalFeature(const std::string& name) const
+{
+	return any([name](auto x) { return x.value == name || x.value == "global." + name; });
+}
