+++ es-app/src/SystemData.h
@@ -15,47 +15,13 @@
 #include "FileFilterIndex.h"
 #include "KeyboardMapping.h"
 #include "math/Vector2f.h"
-#include "utils/VectorEx.h"
+#include "CustomFeatures.h"
 
 class FileData;
 class FolderData;
 class ThemeData;
 class Window;
 class SaveStateRepository;
-
-struct CustomFeatureChoice
-{
-	std::string name;
-	std::string value;
-};
-
-struct CustomFeature
-{
-	CustomFeature()
-	{
-		order = 0;
-	}
-
-	std::string name;
-	std::string value;
-	std::string description;
-	std::string submenu;
-	std::string preset;
-
-	std::string group;
-	int order;
-
-	std::vector<CustomFeatureChoice> choices;
-};
-
-class CustomFeatures : public VectorEx<CustomFeature>
-{
-public:
-	void sort();
-
-	bool hasFeature(const std::string& name) const;
-	bool hasGlobalFeature(const std::string& name) const;
-};
 
 struct GameCountInfo
 {
@@ -67,91 +33,6 @@ struct GameCountInfo
 	int gamesPlayed;
 	std::string mostPlayed;
 	std::string lastPlayedDate;
-};
-
-class EmulatorFeatures
-{
-public:
-	enum Features
-	{
-		none = 0,
-		ratio = 1,
-		rewind = 2,
-		smooth = 4,
-		shaders = 8,
-		pixel_perfect = 16,
-		decoration = 32,
-		latency_reduction = 64,
-		game_translation = 128,
-		autosave = 256,
-		netplay = 512,
-		fullboot = 1024,
-		emulated_wiimotes = 2048,
-		screen_layout = 4096,
-		internal_resolution = 8192,
-		videomode = 16384,
-		colorization = 32768,
-		padTokeyboard = 65536,
-		cheevos = 131072,
-		autocontrollers = 262144,
-
-		all = 0x0FFFFFFF
-	};
-
-	static Features parseFeatures(const std::string features);
-};
-
-struct SystemFeature
-{
-	SystemFeature()
-	{
-		features = EmulatorFeatures::Features::none;
-	}
-
-	std::string name;
-	EmulatorFeatures::Features features;
-	CustomFeatures customFeatures;
-};
-
-struct CoreData
-{
-	CoreData() 
-	{ 
-		netplay = false; 
-		isDefault = false;
-		features = EmulatorFeatures::Features::none;
-	}
-
-	std::string name;
-	bool netplay;
-	bool isDefault;
-	
-	std::string customCommandLine;
-	CustomFeatures customFeatures;
-	std::vector<std::string> incompatibleExtensions;
-
-	EmulatorFeatures::Features features;
-
-	std::vector<SystemFeature> systemFeatures;
-};
-
-struct EmulatorData
-{
-	EmulatorData()
-	{
-		features = EmulatorFeatures::Features::none;
-	}
-
-	std::string name;	
-	std::vector<CoreData> cores;
-
-	std::string customCommandLine;
-	CustomFeatures customFeatures;
-	std::vector<std::string> incompatibleExtensions;
-
-	EmulatorFeatures::Features features;
-
-	std::vector<SystemFeature> systemFeatures;
 };
 
 struct SystemMetadata
@@ -187,9 +68,6 @@ public:
 	static SystemData* getSystem(const std::string name);
 	static SystemData* getFirstVisibleSystem();
 
-	static std::map<std::string, EmulatorData> es_features;
-	static bool es_features_loaded;
-
 	inline FolderData* getRootFolder() const { return mRootFolder; };
 	inline const std::string& getName() const { return mMetadata.name; }
 	inline const std::string& getFullName() const { return mMetadata.fullName; }
@@ -218,12 +96,8 @@ public:
 	static bool loadConfig(Window* window = nullptr); //Load the system config file at getConfigPath(). Returns true if no errors were encountered. An example will be written if the file doesn't exist.
 	static void writeExampleConfig(const std::string& path);
 	static std::string getConfigPath(bool forWrite); // if forWrite, will only return ~/.emulationstation/es_systems.cfg, never /etc/emulationstation/es_systems.cfg
-
-	static bool loadEsFeaturesFile();
 	
 	bool loadFeatures();
-
-	static CustomFeatures loadCustomFeatures(pugi::xml_node node);
 
 	static std::vector<SystemData*> sSystemVector;
 
@@ -320,9 +194,6 @@ public:
 
 	bool shouldExtractHashesFromArchives();
 
-	static CustomFeatures mSharedFeatures;
-	static CustomFeatures mGlobalFeatures;
-
 	bool getShowFilenames();
 	bool getShowParentFolder();
 	bool getShowFavoritesFirst();
@@ -339,7 +210,6 @@ public:
 private:
 	std::string getKeyboardMappingFilePath();
 	static void createGroupedSystems();
-	static bool sortCustomFeatures(CustomFeature& feat1, CustomFeature& feat2);
 
 	size_t mGameListHash;
 
@@ -354,14 +224,6 @@ private:
 	SystemEnvironmentData* mEnvData;
 	std::shared_ptr<ThemeData> mTheme;
 
-	/*
-	std::string mName;
-	std::string mFullName;
-	std::string mThemeFolder;
-	std::string mManufacturer;
-	std::string mReleaseYear;
-	std::string mHardwareType;
-	*/
 	void populateFolder(FolderData* folder, std::unordered_map<std::string, FileData*>& fileMap);
 	void indexAllGameFilters(const FolderData* folder);
 	void setIsGameSystemStatus();
